@He-Pin 2023-07-30
before:
```
[info] Benchmark                                                         Mode  Cnt         Score        Error   Units
[info] FlatMapConcatBenchmark.internalSingleSource                      thrpt   10   9111519.387 锟斤拷 183996.447   ops/s
[info] FlatMapConcatBenchmark.mapBaseline                               thrpt   10  20336793.367 锟斤拷 887065.039   ops/s
[info] FlatMapConcatBenchmark.oneElementList                            thrpt   10    371981.248 锟斤拷   3765.397   ops/s
[info] FlatMapConcatBenchmark.sourceDotSingle                           thrpt   10   4578547.437 锟斤拷 479604.201   ops/s
```

after:
```
info] Benchmark                                                         Mode  Cnt          Score            Error   Units
[info] FlatMapConcatBenchmark.internalSingleSource                      thrpt   10    9027659.060 锟斤拷     136614.084   ops/s
[info] FlatMapConcatBenchmark.mapBaseline                               thrpt   10   20947850.178 锟斤拷    1039800.021   ops/s
[info] FlatMapConcatBenchmark.oneElementList                            thrpt   10    4451591.431 锟斤拷      55815.355   ops/s
[info] FlatMapConcatBenchmark.sourceDotSingle                           thrpt   10    4547444.612 锟斤拷     136653.128   ops/s
```