@GreyPlane 2023-07-31
I wonder if it is possible to write a Buffer.peek(n: Int) which peek from current to n, and by that, is it easy to impl Merge via this, just change `inflightSources.peek() eq self` to check through the range from 0 to `breadth - 1`.