@He-Pin 2023-07-31
In rxjava. There is a concatMap too.