@He-Pin 2023-07-31
Just released should add another parameter `earge` too. When true will complete / fail the stream when upstream fail.