@gael-ft 2023-08-01
> Seems should add another parameter eager too. When true will complete / fail the stream when upstream fail.

Not obvious to me.
Goal `flatMapConcat` is to keep elements ordered.
If ordering is not a concern, you could just use `merge`.

So I can't see a scenario where you would like to continue after one of the upstream sources cancelled when using this operator.

That being said, if you feel that could be useful, let's go 😉