@He-Pin 2023-08-01
@gael-ft that's true. but I saw your PR and you can return `Source.failed(...)` sometime. and for the `eager` I mean `eagerError`.
If when materilization, the upstream is `Source.failed` would I failed the stream ASAP when the `parallesm > 1`?

And for most case, I think we should make the `eagerError = false`, even there is an inflight `Source.failed(...)`.

I was wantted to optimize for `Source.future` and `Source.failed`, but in which way the Source can be already failed. and there are `Flux.concatMapDelayError` I think that should be the default behavior?