@gael-ft 2023-08-01
@He-Pin I might be a bit too much use case centric (i.e. S3 object upload where we don't want any missing part).
 
> If when materilization, the upstream is Source.failed would I failed the stream ASAP when the parallesm > 1?

So this question can probably be a valid question in some cases.
 
But regarding the default value of `eager` I think it should be intentional to continue processing after any error occured.
As with `Flux`, using `flatMapSequentialDelayError` is intentional vs `flatMapSequential`.

In my understanding of `flatMapConcat` it guarantees that the element you process will be in order.
If for some reason you can afford skipping element N+1 and process directly N+2 that should be intentional.

Once again, just my opinion, but as long as this is parameter, users can use it as they want.