@He-Pin 2023-08-01
@gael-ft Yes, let's just keep it simple for now. And I saw the `FailedSource` which will failStage when `preStartI()`. Not sure should that be moved to the `onPull()` block. 
And the `FutureSource` only trigger the failStage after be pulled.