@He-Pin 2024-01-08
@johanandren How about I spliting the optimization to a dedicated PR from this one, I think this PR is too big